Refactor: 하나의 컴포넌트에서만 사용하는 utils함수 해당 파일로 이동
