+++ src/components/myPage/Statistics/index.js
@@ -3,4 +3,4 @@ export { default as StatsWrapper } from './StatsWrapper';
 export { default as StatsByProvider } from './StatsByProvider';
 export { default as StatsByMonthly } from './StatsByMonthly';
 // export { default as StatsByGenre } from './StatsByGenre';
-export { default as StatsByGenre } from '../StatsByGenre';
+export { default as StatsByGenre } from './StatsByGenre';
