--- src/utils/getProvidersIdsByList.js
@@ -1,3 +1,0 @@
-const getProvidersIdsByList = list => list?.flatMap(item => item.id);
-
-export default getProvidersIdsByList;
