+++ src/utils/index.js
@@ -1,11 +1,3 @@
-export { default as getProvidersIdsByList } from './getProvidersIdsByList';
-export { default as getProviderIdsByProviderDatas } from './getProviderIdsByProviderDatas';
-export { default as getProviderArray } from './getProviderArray';
-
-export { default as getNewSubscribeList } from './getNewSubscribeList';
-
 export { default as generateInitial } from './generateInitial';
-
-export { default as formatDate } from './formatDate';
 
 export { default as goToTop } from './goToTop';
